# swagger.yaml
openapi: 3.0.3
info:
  title: Branded UK E-Commerce API
  description: |
    # Overview
    This API provides endpoints for the Branded UK e-commerce platform.
    
    ## Base URL
    **Production**: `https://api.brandeduk.com`
    
    ## Authentication
    No authentication required. Public API.
    
    ## Endpoint Categories
    
    ### Products (Main)
    - `GET /api/products` - Main product listing with combined filters
    - `GET /api/products/filters` - Product type counts for filtering
    - `GET /api/products/types` - All product types with counts
    - `GET /api/products/{code}` - Single product details
    - `GET /api/products/{code}/pricing` - Get pricing information for a style code
    
    ### Contact & Quotes
    - `POST /api/contact` - Submit contact form inquiry
    - `POST /api/quotes` - Submit quote request with optional logo file uploads
    
    ### Filter Endpoints (Testing & Database Exploration)
    Each filter type has TWO endpoints:
    1. **List all values**: Returns all options from that database table
    2. **Get products by slug**: Returns products matching a specific filter value
    
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: https://api.brandeduk.com
    description: Production Server
  - url: http://localhost:3004
    description: Local Development
...
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get product list (main combined filter endpoint)
      description: |
        Returns a paginated list of products with optional filtering and search.
        This is the main endpoint that supports all filters combined.
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page (max 200)
          schema:
            type: integer
            default: 24
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: priceMin
          in: query
          schema:
            type: number
        - name: priceMax
          in: query
          schema:
            type: number
        - name: productType
          in: query
          description: Product type slug(s)
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: ageGroup
          in: query
          schema:
            type: string
        - name: sleeve
          in: query
          schema:
            type: string
        - name: neckline
          in: query
          schema:
            type: string
        - name: fabric
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: colour
          in: query
          schema:
            type: string
        - name: primaryColour
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: weight
          in: query
          schema:
            type: string
        - name: fit
          in: query
          schema:
            type: string
        - name: sector
          in: query
          schema:
            type: string
        - name: sport
          in: query
          schema:
            type: string
        - name: effect
          in: query
          schema:
            type: string
        - name: accreditations
          in: query
          schema:
            type: string
        - name: brand
          in: query
          description: Brand slug(s) or name(s) - accepts multiple values
          schema:
            type: string
          example: "awdis"
        - name: brands
          in: query
          description: Alias for brand parameter (accepts multiple values)
          schema:
            type: string
        - name: colourShade
          in: query
          schema:
            type: string
        - name: sort
          in: query
          description: |
            Sort option. Available values:
            - `newest` - Sort by creation date (newest first)
            - `best` - Best sellers (uses newest as proxy)
            - `brand-az` - Sort by brand name A-Z
            - `brand-za` - Sort by brand name Z-A
            - `code-az` - Sort by product code A-Z
            - `code-za` - Sort by product code Z-A
            - `price-lh` - Sort by price low to high
            - `price-hl` - Sort by price high to low
            - `price` - Sort by price (use with order parameter)
            - `name` - Sort by product name (use with order parameter)
          schema:
            type: string
            enum: [newest, best, brand-az, brand-za, code-az, code-za, price-lh, price-hl, price, price-low, price-high, name]
            default: newest
            default: newest
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  priceRange:
                    type: object
                    properties:
                      min:
                        type: number
                      max:
                        type: number
  /api/products/filters:
    get:
      tags:
        - Products
      summary: Get filter aggregations (counts)
      description: Returns counts for all filter types based on current filters
      operationId: getProductFilters
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: productType
          in: query
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: ageGroup
          in: query
          schema:
            type: string
        - name: sleeve
          in: query
          schema:
            type: string
        - name: neckline
          in: query
          schema:
            type: string
        - name: fabric
          in: query
          schema:
            type: string
        - name: size
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  filters:
                    type: object
                    properties:
                      gender:
                        type: object
                        description: Map of gender slug to count
                      ageGroup:
                        type: object
                        description: Map of age group slug to count
                      sleeve:
                        type: object
                        description: Map of sleeve slug to count
                      neckline:
                        type: object
                        description: Map of neckline slug to count
                      fabric:
                        type: object
                        description: Map of fabric slug to count
                      size:
                        type: object
                        description: Map of size slug to count
                      feature:
                        type: object
                        description: Map of feature slug to count
                      tag:
                        type: object
                        description: Map of tag slug to count
...
  /api/products/{code}:
    get:
      tags:
        - Products
      summary: Get product details by code
      operationId: getProductDetail
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
          example: "JC001"
      responses:
        '200':
          description: Successful response
        '404':
          description: Product not found
  /api/products/{code}/related:
    get:
      tags:
        - Products
      summary: Get related products
      description: |
        Returns products related to the specified product.
        Priority: Same brand AND same product type first, then same product type from other brands.
      operationId: getRelatedProducts
      parameters:
        - name: code
          in: path
          required: true
          description: Product style code
          schema:
            type: string
          example: "JC001"
        - name: limit
          in: query
          description: Maximum number of related products to return
          schema:
            type: integer
            default: 12
          example: 12
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  currentProduct:
                    type: object
                    properties:
                      code:
                        type: string
                      brand:
                        type: string
                      productType:
                        type: string
                  related:
                    type: array
                    items:
                      type: object
                      properties:
                        code:
                          type: string
                        name:
                          type: string
                        price:
                          type: number
                        image:
                          type: string
                        brand:
                          type: string
                        product_type:
                          type: string
                  total:
                    type: integer
                  sameBrandAndType:
                    type: integer
                    description: Number of products from same brand AND type
                  sameTypeOnly:
                    type: integer
                    description: Number of products from same type only (different brand)
        '404':
          description: Product not found
  /api/products/{code}/pricing:
    get:
      tags:
        - Products
      summary: Get pricing information for a style code
      description: |
        Returns pricing breakdown for a product including:
        - carton_price (or single_price if carton_price is 0/null)
        - markup_percent (the pricing rule percentage applied)
        - sell_price (the calculated selling price)
        - pricing_rule details (price range and description)
      operationId: getProductPricing
      parameters:
        - name: code
          in: path
          required: true
          description: Product style code
          schema:
            type: string
          example: "JC001"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  style_code:
                    type: string
                    example: "JC001"
                  carton_price:
                    type: number
                    description: Carton price (or single_price if carton_price is 0/null)
                    example: 10.50
                  single_price:
                    type: number
                    description: Single unit price
                    example: 10.50
                  markup_percent:
                    type: number
                    nullable: true
                    description: Markup percentage applied from pricing rules
                    example: 127.00
                  sell_price:
                    type: number
                    nullable: true
                    description: Calculated selling price after markup
                    example: 23.84
                  pricing_rule:
                    type: object
                    nullable: true
                    description: Details of the pricing rule applied
                    properties:
                      from_price:
                        type: number
                        example: 10.00
                      to_price:
                        type: number
                        example: 14.99
                      markup_percent:
                        type: number
                        example: 127.00
                      description:
                        type: string
                        example: "127% markup for £10.00-£14.99 (reduced by 5%)"
        '404':
          description: Product not found or has no pricing information
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Product not found or has no pricing information"
                  style_code:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # PRICING ENDPOINTS
  # ============================================================================
...
  /api/filters/product-types/{slug}/products:
    get:
      tags:
        - Product Types
      summary: Get products by product type slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: "t-shirts"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

  # ============================================================================
  # GENDER ENDPOINTS
  # ============================================================================
...
  /api/filters/genders/{slug}/products:
    get:
      tags:
        - Genders
      summary: Get products by gender slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: "male"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

  # ============================================================================
  # AGE GROUP ENDPOINTS
  # ============================================================================
...
  /api/filters/brands/{slug}/products:
    get:
      tags:
        - Brands
      summary: Get products by brand slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
          example: "awdis"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 24
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...